LayoutTests/imported/w3c/web-platform-tests/intersection-observer/containing-block-expected.txt


PASS IntersectionObserver should only report intersections if root is a containing block ancestor of target. 
FAIL In containing block and intersecting. assert_equals: entries[0].boundingClientRect.left expected 58 but got 0
FAIL In containing block and not intersecting. assert_equals: entries.length expected 2 but got 1
FAIL Not in containing block and intersecting. assert_equals: entries.length expected 2 but got 1
FAIL Not in containing block and not intersecting. assert_equals: entries.length expected 2 but got 1
